% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arbin_Analysis.R
\name{Arbin.VP}
\alias{Arbin.VP}
\alias{Arbin.CAPA}
\title{Arbin_analysis}
\usage{
Arbin.VP(raw, AMmass, cellType, cycles = c(1, 5, 10))

Arbin.CAPA(raw, AMmass, cellType)
}
\arguments{
\item{raw}{raw data file}

\item{AMmass}{- object of Biologic.CAPA}

\item{cellType}{- object of Biologic.CAPA}

\item{cycles}{selected cycles to extract for voltage profiles}
}
\value{
returns data.frames (for Arbin.VP as list)
}
\description{
Analysis Arbin data to produce data.frames for voltage profiles of selected cycles
and capacity (vs. cycle number) data
}
\details{
Extract voltage profiles for selected cycles

Arbin.CAPA extracts cycling data from galvanostatic cycling experiments
}
\examples{
\dontrun{
capa <- Arbin.CAPA()

VP <- Arbin.VP()
}
}
