% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Read0r2.R
\name{BCSraw}
\alias{BCSraw}
\title{Read experimental data}
\usage{
BCSraw(dir, filename)
}
\arguments{
\item{filename}{}
}
\value{

}
\description{
Imports experimental data from data folders, depending on data format/instrument.
Experimental data is converted into a uniform format for further processing
}
