% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Biologic_Analysis.R
\name{Biologic.CAP}
\alias{Biologic.CAP}
\title{Analysis of Biologic potentiostat data}
\usage{
Biologic.CAP(raw, AM.mass, type)
}
\arguments{
\item{raw}{- object of Biologic.CAP}

\item{AM.mass}{- object of Biologic.CAP}

\item{type}{- object of Biologic.CAP}
}
\value{
capacity table
}
\description{
Analysis of Biologic potentiostat data
}
\examples{

}
