% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Evaluat0r.R
\name{BiologicEvaluat0r}
\alias{BiologicEvaluat0r}
\title{Starts evaluation of expeirmental data}
\usage{
BiologicEvaluat0r(raw, AMmass, cellType, cycles, cccv, warningsLOG)
}
\arguments{
\item{warningsLOG}{}
}
\value{

}
\description{
Depending on instrument different analysis scripts are needed
}
