% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customThemes.R
\name{customTheme}
\alias{customTheme}
\title{customTheme}
\usage{
customTheme()
}
\value{
customTheme for ggplot2
}
\description{
creates a custom theme for plotting data, used in report.Rmd
}
\examples{
p <- ggplot(mtcars) +
     geom_point(aes(x=mpg, y=cyl), color='red', size=4) +
     customTheme()

p <- p + customTheme()
}
