% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Evaluat0r.R
\name{BiologicEvaluat0r}
\alias{BiologicEvaluat0r}
\alias{ArbinEvaluat0r}
\title{Evaluat0r}
\usage{
BiologicEvaluat0r(raw, AMmass, cellType, cycles, warningsLOG)

ArbinEvaluat0r(raw, AMmass, cellType, cycles)
}
\arguments{
\item{raw}{raw data}

\item{AMmass}{active material mass in mg}

\item{cellType}{cell configuration (halfcell, fullcell, etc.)}

\item{cycles}{cycles to be included in voltage profile analysis}

\item{warningsLOG}{a log book for warning messages}
}
\value{
returns a list cycDat with analysed data
}
\description{
Starts evaluation of experimental data.
Depending on instrument different analysis scripts are needed
}
\details{
Evaluat0r for Biologic Instruments

Evaluat0r for Arbin Instruments
}
\examples{
\dontrun{
BiologicEvaluat0r()
}
}
