% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exampleData}
\alias{exampleData}
\title{Example data for galvanostatic cell cycling.}
\format{
A data frame with 53940 rows and 10 variables:
\describe{
  \item{cycle.number}{cycle number}
  \item{Ns}{sequence counter}
  \item{time.s}{experiment time}
  \item{X.I..mA}{applied current}
  \item{Ecell.V}{cell potential}
  \item{Q.discharge.mA.h}{negative charge passed}
  \item{Q.charge.mA.h}{positive charge passed}
  \item{dq.mAh}{charge increments}
  \item{X.Q.Q0..mA.h}{Difference between Q and Q0}
  \item{Capacity.mA.h}{capacity}
  \item{Energy.charge.W.h}{energy on charge sequences}
  \item{Energy.discharge.W.h}{energy on discharge sequences}
  \item{X}{empty column}
}
}
\source{
galvanostatic cycling experiment of a graphite-potassium half cell measured on a Biologic BCS instrument
}
\usage{
exampleData
}
\description{
A typical dataset for a galvanostatic cycling experiments of batteries.
}
\keyword{datasets}
