% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exampleMeta}
\alias{exampleMeta}
\title{Example meta data file of a battery cell.}
\format{
A data frame with 53940 rows and 10 variables:
\describe{
  \item{Identifier}{identifier of experiment}
  \item{sample name}{name of the file *without* file extension}
  \item{instrument}{type of potentiostat used, with currently three options:
  *"Biologic BCS", "Biologic VMP"* (= also for MPG and VSP instruments), *"Arbin"*}
  \item{cell config}{cell setup with four options:
  *"halfcell-cathode", "fullcell", "halfcell-anode", "LiS"*;
  **Note that spelling need to be exact.**}
  \item{AM loading}{absolut active material content of working electrode}
}
}
\source{
galvanostatic cycling experiment of a graphite-potassium half cell measured on a Biologic BCS instrument
}
\usage{
exampleMeta
}
\description{
Specifics of meta data file.
}
\keyword{datasets}
