% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cccv.R
\name{\%>\%}
\alias{\%>\%}
\alias{Biologic.CCCV}
\title{CCCV.Analysis}
\usage{
Biologic.CCCV(raw, AMmass, cellType)
}
\arguments{
\item{raw}{raw data}

\item{AMmass}{active material mass in mg}

\item{cellType}{cell configuration (halfcell, fullcell, etc.)}
}
\value{
revised capacity/cccv table
}
\description{
Biologic.CAP Evaluates CCCV-steps in galvanostatic cycling data
}
