% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Direct0r.R
\name{metaDir}
\alias{metaDir}
\title{Direct0r}
\usage{
metaDir()
}
\arguments{
\item{cycles}{cycles to be extracted for voltage profiles}

\item{cccv}{was cccv cycling conducted? TRUE/FALSE}
}
\value{
returns a data.frame with metadata on samples and an input & output directory
}
\description{
Locate folder with meta data and raw data
}
\examples{

}
