% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot0r.R
\name{plotGC}
\alias{plotGC}
\alias{plotIRdrop}
\alias{plotIntR}
\alias{plotVP}
\alias{plotdQdV}
\alias{plotdVdQ}
\title{plot0r}
\usage{
plotGC(cell)

plotIRdrop(cell)

plotIntR(cell, minmax = c(0, 1000, 0, 1000))

plotVP(cell)

plotdQdV(cell)

plotdVdQ(cell)
}
\arguments{
\item{cell}{data.frame with galvanostatic cycling information (Voltage Profiles or Cycling Data)}

\item{minmax}{y-axis bounds}
}
\value{
returns graphs generated from expeirmental data for report.Rmd
}
\description{
Plotting Script for RMarkdown Report
}
\details{
plots of galvanostatic cycling (GC) experiments (Q vs. CycNr & CE vs. CycNr)

plotIRdrop plots IR drop versus cycle number (charge)

plotIntR plots internal resistance versus cycle number (charge)

plotVP plots voltage profile in different presentations

plotdQdV plots differential capacity plots (dQ/dV) - uses a moving average to smooth the data!

plotdVdQ plots differential voltage plots (dV/dQ) - uses a moving average to smooth the data!
}
\examples{
\dontrun{
capa <- Arbin.CAP()
pCap <- plotCapa(capa)
}

}
