% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot0r.R
\name{plotCapReport}
\alias{plotCapReport}
\alias{plotIRdrop}
\alias{plotVPloop}
\alias{plotVPlin}
\alias{plotVPsplit}
\title{plot0r}
\usage{
plotCapReport(capacity)

plotIRdrop(capacity)

plotVPloop(vp.dat)

plotVPlin(vp.dat)

plotVPsplit(vp.dat, cell)
}
\arguments{
\item{capacity}{data.frame with galvanostatic cycling information}

\item{vp.dat}{data.frame with voltage profile information on selected cycles}

\item{cell}{cell configuration of experiment (halfcell-anode, halfcell-cathode, fullcell, LiS)}
}
\value{
returns graphs generated from expeirmental data for report.Rmd
}
\description{
Plotting Script for RMarkdown Report
}
\section{Functions}{
\itemize{
\item \code{plotIRdrop}: plots IR drop versus cycle number

\item \code{plotVPloop}: plots voltage profile vs Qloop

\item \code{plotVPlin}: plots voltage profile vs Q

\item \code{plotVPsplit}: plots voltage profile vs Q
}}

