% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Process0r.R
\name{process0r}
\alias{process0r}
\title{Process dataset}
\usage{
process0r(cycles = c(seq(0, 100, 10)), cccv = FALSE)
}
\arguments{
\item{cycles}{cycles to be extracted for voltage profiles}

\item{cccv}{was cccv cycling conducted? TRUE/FALSE}
}
\value{

}
\description{
Initiates import and analysis of a curated dataset
}
