% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renderReport.R
\name{renderReport}
\alias{renderReport}
\title{renderReport}
\usage{
renderReport(sample = exp, RmdFile = "report-new.Rmd")
}
\arguments{
\item{sample}{cell data of ONE cell (list); can only process one cell at a time}

\item{RmdFile}{name of RMarkdown Report template}

\item{exp}{sample info}
}
\value{
initiates data report function report-new.Rmd. Style of RMarkdown Report could be adjusted as needed: edit RmdFile in renderReport(RmdFile = 'filename') and add RMarkdown file to inst > rmd folder of package
}
\description{
Interface to plot data and pass it to report.Rdm to
generate data report
}
\examples{
\dontrun{
exp <- list("metadata" <- base.info,
            "capacity" <- data.frame,
            "VoltageProfiles" <- list(df1, df2, ...)
            )
renderReport(exp)
}

}
