% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Report0r.R
\name{report0r}
\alias{report0r}
\title{Report0r}
\usage{
report0r(
  cycles = c(0, 1, 4, seq(9, 199, 10)),
  CCCV = FALSE,
  htmlReport = FALSE,
  exportCap = TRUE
)
}
\arguments{
\item{cycles}{cycle numbers to be extracted to plot voltage profiles}

\item{CCCV}{should a CCCV analysis be performed? (TRUE/FALSE)}

\item{htmlReport}{create a html report? TRUE/FALSE}

\item{exportCap}{export data.frames to .txt for further processing? TRUE/FALSE}
}
\value{
no return; saves a html report, .txt files of analysed data or both
}
\description{
initiate data processing with:
report0r(htmlReport = TRUE, exportCap = TRUE)
}
\examples{
\dontrun{
 dat <- report0r(htmlReport = TRUE, exportCap = TRUE)
 }
}
