% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportGenerat0r.R
\name{reportGenerat0r}
\alias{reportGenerat0r}
\title{reportGenerat0r}
\usage{
reportGenerat0r(exp)
}
\arguments{
\item{exp}{sample info}
}
\value{
initiates data report function report.Rmd and generates plots for report using plot0r.R
}
\description{
Interface to plot data and pass it to report.Rdm to
generate data report
}
\details{
Produce plots for HTML report
}
\examples{
\dontrun{
exp <- list("metadata" <- base.info,
            "capacity" <- data.frame,
            "VoltageProfiles" <- list(df1, df2, ...),
            "CCCV" <- data.frame)
report <- reportGenerat0r(exp)
}

}
