% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme1.R
\name{theme1}
\alias{theme1}
\title{theme1}
\usage{
theme1()
}
\value{
theme1 for ggplot2
}
\description{
creates a custom theme for plotting data, used in report.Rmd
}
\examples{
p <- ggplot(mtcars) +
     geom_point(aes(x=mpg, y=cyl), color='red', size=4) +
     theme1()

p <- p + theme1()
}
