% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme2.R
\name{theme2}
\alias{theme2}
\title{theme2}
\usage{
theme2()
}
\value{
theme2 for ggplot2
}
\description{
creates a custom theme for plotting data, used in report.Rmd
}
\examples{
p <- ggplot(mtcars) +
     geom_point(aes(x=mpg, y=cyl), color='red', size=4) +
     theme2()

p <- p + theme2()
}
