% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TXTExport0r.R
\name{SaveStatsToOrigin}
\alias{SaveStatsToOrigin}
\alias{SaveCCCVToOrigin}
\alias{SaveVPToOrigin}
\title{TXTExport0r}
\usage{
SaveStatsToOrigin(exp)

SaveCCCVToOrigin(exp)

SaveVPToOrigin(exp)
}
\arguments{
\item{exp}{list containing all sample information}
}
\value{
no return; writes a .txt file with respective data into outdir folder
}
\description{
exports analysed data from list 'processedData', i.e. exp = processedData[[sampleID]]
}
\details{
Save Stats ("cycling data" --> Q, CE,..)

Save CCCV analysis data

Save VPs for selected cycles
}
\examples{
\dontrun{
l <- process0r()
exp <- l[[1]]
SaveStatsToOrigin(exp)
}

}
