% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Biologic_Analysis.R
\name{Biologic.CAPA}
\alias{Biologic.CAPA}
\alias{Biologic.VP}
\title{Biologic_Analysis}
\usage{
Biologic.CAPA(raw, AMmass, cellType)

Biologic.VP(raw, AMmass, cycles, cellType)
}
\arguments{
\item{raw}{- object of Biologic.CAPA}

\item{AMmass}{- object of Biologic.CAPA}

\item{cellType}{- object of Biologic.CAPA}

\item{cycles}{- selected cycles to extract for voltage profiles}
}
\value{
capacity table
}
\description{
Analysis of Biologic potentiostat data
}
\details{
Extract Cycling Data for Biologic Instruments

Biologic.VP extracts voltage profiles for selected cycles from cycling data
}
\examples{
\dontrun{
capa <- Biologic.CAP()

VP <- Biologic.VP()
}
}
