% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Read0r2.R, R/Evaluat0r.R, R/Arbin_Analysis.R,
%   R/Biologic_Analysis.R
\name{\%>\%}
\alias{\%>\%}
\alias{BCSraw}
\alias{BiologicEvaluat0r}
\alias{Arbin.VP}
\alias{Biologic.CAP}
\title{Read experimental data}
\usage{
BCSraw(dir, filename)

BiologicEvaluat0r(raw, AMmass, cellType, cycles, cccv, warningsLOG)

Arbin.VP(raw, cycles)

Biologic.CAP(raw, AM.mass, type)
}
\arguments{
\item{dir}{directory to raw data folder}

\item{filename}{sample name}

\item{raw}{- object of Biologic.CAP}

\item{AMmass}{active material mass in mg}

\item{cellType}{cell configuration (halfcell, fullcell, etc.)}

\item{cycles}{}

\item{cccv}{perform CCCV step analysis? TRUE/FALSE}

\item{warningsLOG}{a log book for warning messages}

\item{AM.mass}{- object of Biologic.CAP}

\item{type}{- object of Biologic.CAP}
}
\value{


returns a list cycDat with analysed data

returns data.frames (for Arbin.VP as list)

capacity table
}
\description{
Imports experimental data from data folders, depending on data format/instrument.
Experimental data is converted into a uniform format for further processing

Depending on instrument different analysis scripts are needed

Analysis Arbin data to produce data.frames for voltage profiles of selected cycles
and capacity (vs. cycle number) data
}
\examples{

}
