% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Process0r.R
\name{metaDir}
\alias{metaDir}
\title{Process dataset}
\usage{
metaDir()
}
\arguments{
\item{cycles}{cycles to be extracted for voltage profiles}

\item{cccv}{was cccv cycling conducted? TRUE/FALSE}
}
\value{

}
\description{
Initiates import and analysis of a curated dataset
}
