% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Direct0r.R
\name{metaDir}
\alias{metaDir}
\title{Direct0r}
\usage{
metaDir()
}
\value{
returns a data.frame with metadata on samples and an input & output directory
}
\description{
Locate folder with meta data and raw data
}
\examples{
\dontrun{
meta <- metaDir()
}
}
