% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot0r.R
\name{plotCapa}
\alias{plotCapa}
\alias{plotCE}
\alias{plotIRdropCH}
\alias{plotIRdropDC}
\alias{plotIntRCH}
\alias{plotIntRDC}
\alias{plotVPloop}
\alias{plotVPlin}
\alias{plotVPsplitDC}
\alias{plotVPsplitCH}
\alias{plotdQdV}
\alias{plotdVdQ}
\title{plot0r}
\usage{
plotCapa(capacity)

plotCE(capacity)

plotIRdropCH(capacity, minmax)

plotIRdropDC(capacity, minmax)

plotIntRCH(capacity, minmax = c(0, 1000, 0, 1000))

plotIntRDC(capacity, minmax = c(0, 1000, 0, 1000))

plotVPloop(vp.dat)

plotVPlin(vp.dat)

plotVPsplitDC(dc, min.dc.y, max.dc.y)

plotVPsplitCH(ch, min.ch.y, max.ch.y)

plotdQdV(vp.dat)

plotdVdQ(vp.dat)
}
\arguments{
\item{capacity}{data.frame with galvanostatic cycling information}

\item{minmax}{axis range}

\item{vp.dat}{voltage profile information}

\item{dc}{data.frame with discharge data for selected cycles}

\item{min.dc.y}{y-axis limit - min}

\item{max.dc.y}{y-axis limit - max}

\item{ch}{data.frame with charge data for selected cycles}

\item{min.ch.y}{y-axis limit - min}

\item{max.ch.y}{y-axis limit - max}
}
\value{
returns graphs generated from expeirmental data for report.Rmd
}
\description{
Plotting Script for RMarkdown Report
}
\details{
plotCapa plots coulombic efficiency versus cycle number

plotCE plots coulombic efficiency versus cycle number

plotIRdropCH plots IR drop versus cycle number (charge)

plotIRdropDC plots IR drop versus cycle number (discharge)

plotIntRCH plots internal resistance versus cycle number (charge)

plotIntRDC plots internal resistance versus cycle number (discharge)

plotVPloop plots voltage profile vs Qloop

plotVPlin plots voltage profile vs Q

plotVPsplitDC plots voltage profile vs Q (discharge)

plotVPsplitCH plots voltage profile vs Q (charge)

plotVPloop plots voltage profile vs Qloop

plotVPloop plots voltage profile vs Qloop
}
\examples{
\dontrun{
capa <- Arbin.CAP()
pCap <- plotCapa(capa)
}
}
