% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Process0r.R
\name{process0r}
\alias{process0r}
\title{Process0r}
\usage{
process0r(cccv = FALSE, cycles = c(seq(0, 99, 10)))
}
\arguments{
\item{cccv}{was cccv cycling conducted? TRUE/FALSE}

\item{cycles}{cycles to be extracted for voltage profiles}
}
\value{
returns a summary of all sample data, including metadata and raw file.
sampleSUMMARY is a list of one or more samples stated in meta.csv file (=cell log).
}
\description{
Initiates import and analysis of a curated dataset
}
\examples{
\dontrun{
l <- process0r()
}
}
