% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Report0r.R
\name{report0r}
\alias{report0r}
\title{Report0r}
\usage{
report0r(htmlReport = FALSE, exportCap = TRUE)
}
\arguments{
\item{htmlReport}{create a html report? TRUE/FALSE}

\item{exportCap}{export data.frames to .txt for further processing? TRUE/FALSE}
}
\value{

}
\description{
initiate data processing with:
analyz0r <- report0r(htmlReport = TRUE, exportCap = TRUE)
}
