% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportGenerat0r.R
\name{reportGenerat0r}
\alias{reportGenerat0r}
\title{reportGenerat0r}
\usage{
reportGenerat0r(exp)
}
\arguments{
\item{exp}{sample info}
}
\value{

}
\description{
Interface to plot data and pass it to report.Rdm to
generate data report
}
